Add option saveReadCorrections.

+++ src/pipelines/canu/CorrectReads.pm
@@ -905,49 +905,53 @@ sub dumpCorrectedReads ($$) {
 
     #  Now that all outputs are (re)written, cleanup the job outputs.
 
+    print STDERR "--\n";
     print STDERR "-- Purging correctReads output after merging to final output file.\n";
+    print STDERR "-- Purging disabled by saveReadCorrections=true.\n"  if (getGlobal("saveReadCorrections") == 1);
 
     my $Nsuccess = 0;
     my $Nerr     = 0;
     my $Nfasta   = 0;
     my $Nlog     = 0;
 
-    open(F, "< $path/corjob.files")             or caExit("can't open '$path/corjob.files' for reading: $!", undef);
-    while (<F>) {
-        chomp;
+    if (getGlobal("saveReadCorrections") != 1) {
+        open(F, "< $path/corjob.files") or caExit("can't open '$path/corjob.files' for reading: $!", undef);
+        while (<F>) {
+            chomp;
 
-        if (m/^(.*)\/correction_outputs\/0*(\d+).fasta$/) {
-            my $ID6 = substr("00000" . $2, -6);
-            my $ID4 = substr("000"   . $2, -4);
-            my $ID0 = $2;
+            if (m/^(.*)\/correction_outputs\/0*(\d+).fasta$/) {
+                my $ID6 = substr("00000" . $2, -6);
+                my $ID4 = substr("000"   . $2, -4);
+                my $ID0 = $2;
 
-            if (-e "$1/correction_outputs/$ID4.dump.success") {
-                $Nsuccess++;
-                unlink "$1/correction_outputs/$ID4.dump.success";
-            }
-            if (-e "$1/correction_outputs/$ID4.err") {
-                $Nerr++;
-                unlink "$1/correction_outputs/$ID4.err";
-            }
-            if (-e "$1/correction_outputs/$ID4.fasta") {
-                $Nfasta++;
-                unlink "$1/correction_outputs/$ID4.fasta";
-            }
+                if (-e "$1/correction_outputs/$ID4.dump.success") {
+                    $Nsuccess++;
+                    unlink "$1/correction_outputs/$ID4.dump.success";
+                }
+                if (-e "$1/correction_outputs/$ID4.err") {
+                    $Nerr++;
+                    unlink "$1/correction_outputs/$ID4.err";
+                }
+                if (-e "$1/correction_outputs/$ID4.fasta") {
+                    $Nfasta++;
+                    unlink "$1/correction_outputs/$ID4.fasta";
+                }
 
-            if (-e "$1/correctReads.$ID6.out") {
-                $Nlog++;
-                unlink "$1/correctReads.$ID6.out";
-            }
-            if (-e "$1/correctReads.$ID0.out") {
-                $Nlog++;
-                unlink "$1/correctReads.$ID0.out";
-            }
+                if (-e "$1/correctReads.$ID6.out") {
+                    $Nlog++;
+                    unlink "$1/correctReads.$ID6.out";
+                }
+                if (-e "$1/correctReads.$ID0.out") {
+                    $Nlog++;
+                    unlink "$1/correctReads.$ID0.out";
+                }
 
-        } else {
-            caExit("unknown correctReads job name '$_'\n", undef);
+            } else {
+                caExit("unknown correctReads job name '$_'\n", undef);
+            }
         }
+        close(F);
     }
-    close(F);
 
     print STDERR "-- Purged $Nsuccess .dump.success sentinels.\n"   if ($Nsuccess > 0);
     print STDERR "-- Purged $Nfasta .fasta outputs.\n"              if ($Nfasta > 0);
